backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.3.0,<3.0.0
numpy>=1.26.0,<3.0.0
pillow>=10.0.0,<12.0.0
opencv-python>=4.8.0,<5.0.0
tensorflow>=2.20.0,<2.21.0
httpx>=0.27.0,<1.0.0
